// src/twitchAPI-types.ts
export interface TwitchToken {
  access_token: string;
  refresh_token?: string;
  expires_in: number;
  scope: string[];
  token_type: string;
}

export interface ClientTwitchExtension {
  api: TwitchAPI;
  auth: {
    api: TwitchToken;
    tmi: TwitchToken;
  };
  banned_users: DBTwitchUser[];
  permitted_users: DBTwitchUser[];
  bot: TwitchUser;
  wordBank: {
    // allowed: string[];
    blocked: string[];
  };
  stream_data: StreamData;
  stream?: TwitchStream;
}

export interface TwitchUser {
  id: string;
  login: string;
  display_name: string;
  type: string;
  broadcaster_type: string;
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  view_count: number;
  email: string;
  created_at: string;
}

export interface DBTwitchUser {
  id: string;
  login: string;
  display_name: string;
  type: string;
  broadcaster_type: string;
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  view_count: number;
  email: string;
  created_at: string;
  banned_date?: string;
  permitted_date?: string;
}

export interface TwitchUsersResponse {
  data: TwitchUser[];
}

export interface TwitchFollow {
  from_id: string;
  from_login: string;
  from_name: string;
  to_id: string;
  to_login: string;
  to_name: string;
  followed_at: string;
}

export interface TwitchFollowsResponse {
  data: TwitchFollow[];
  pagination: {
    cursor?: string;
  };
  total: number;
}

export interface TwitchStream {
  id: string;
  user_id: string;
  user_login: string;
  user_name: string;
  game_id: string;
  game_name: string;
  type: string;
  title: string;
  viewer_count: number;
  started_at: string;
  language: string;
  thumbnail_url: string;
  tag_ids: string[];
  is_mature: boolean;
}
// ...
export interface StreamData {
  raiders: string[];
  subs: string[];
  reSubs: string[];
  gifters: any;
  // seedBank: liveData.seedBank ?? [],
  welcomeList: string[];
  cheerers: string[];
  binxRaids: string[];
}